Fix: #124 - Handle data type mismatch for expiresIn and cNonceExpiresIn in token response

+++ eudi-wallet-oidc-android/src/main/java/com/ewc/eudi_wallet_oidc_android/models/TokenResponse.kt
@@ -6,10 +6,10 @@ data class TokenResponse(
 
     @SerializedName("access_token") var accessToken: String? = null,
     @SerializedName("token_type") var tokenType: String? = null,
-    @SerializedName("expires_in") var expiresIn: Int? = null,
+    @SerializedName("expires_in") var expiresIn: Long? = null,
     @SerializedName("id_token") var idToken: String? = null,
     @SerializedName("c_nonce") var cNonce: String? = null,
-    @SerializedName("c_nonce_expires_in") var cNonceExpiresIn: Int? = null,
+    @SerializedName("c_nonce_expires_in") var cNonceExpiresIn: Long? = null,
     @SerializedName("error") var error: String? = null,
     @SerializedName("error_description") var errorDescription: String? = null
 )
